# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "safehaven",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/comments": {
      "post": {
        "tags": [
          "admin::comments"
        ],
        "operationId": "admin_comment_new",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewComment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Comment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/comments/pending": {
      "get": {
        "tags": [
          "admin::comments"
        ],
        "operationId": "admin_comments_pending",
        "responses": {
          "200": {
            "description": "List of pending comments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListedComment"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/comments/{id}": {
      "get": {
        "tags": [
          "admin::comments"
        ],
        "operationId": "admin_comment_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Comment identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin::comments"
        ],
        "operationId": "admin_comment_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Comment identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateComment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Comment updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin::comments"
        ],
        "operationId": "admin_comment_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Comment identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted successfully"
          },
          "401": {
            "description": "Invalid permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Comment": {
        "type": "object",
        "required": [
          "id",
          "entity_id",
          "author",
          "text",
          "data",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "data": {},
          "entity_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "moderated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "moderated_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "text": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error_code"
        ],
        "properties": {
          "details": {
            "type": "string",
            "nullable": true
          },
          "error_code": {
            "type": "string"
          }
        }
      },
...
      "ListedComment": {
        "type": "object",
        "required": [
          "id",
          "entity_id",
          "entity_display_name",
          "entity_category_id",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "entity_category_id": {
            "type": "string",
            "format": "uuid"
          },
          "entity_display_name": {
            "type": "string"
          },
          "entity_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "NewComment": {
        "type": "object",
        "required": [
          "entity_id",
          "author",
          "text",
          "data"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "data": {},
          "entity_id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string"
          }
        }
      },
...
      "UpdateComment": {
        "type": "object",
        "required": [
          "entity_id",
          "author",
          "text",
          "data"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "data": {},
          "entity_id": {
            "type": "string",
            "format": "uuid"
          },
          "moderated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "moderated_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "text": {
            "type": "string"
          }
        }
      },
...
    }
  }
}